refactor: establish src/design/ as the design system
Dissolve src/web/ — the design system is now a first-class concern:
- src/design/ owns all stateless UI: CSS, components, templates, assets
- src/pages/ owns route entrypoints that pass data into design components
- src/ now has 4 top-level dirs: build/, catalog/, design/, pages/

Each component gets its own subdirectory (forms-lab pattern):
- Per-component styles.css split from the monolithic components.css
- Tag, button, link, select, card, side-nav stubs ready for implementation

Add compositions layer (l-stack, l-cluster, l-sidebar) for layout utilities.
CSS layers: reset, tokens, base, compositions, layout, components, utilities.
Build copies only .css files to dist/design/ via new copyCssTree helper.
