+++ src/design/common/content-page.css
@@ -1,0 +1,3 @@
+.content-page { display: grid; gap: var(--space-4); }
+.content-page h2 { margin-block-start: var(--space-6); }
+.content-page h3 { margin-block-start: var(--space-5); }
