+++ src/design/components/catalog-filter/styles.css
@@ -1,0 +1,22 @@
+catalog-filter { display: block; }
+
+.catalog-filter {
+  margin-block-end: var(--space-5);
+}
+
+.catalog-filter fieldset {
+  border: 1px solid var(--color-surface-alt);
+  border-radius: var(--radius-md);
+  padding: var(--space-4);
+  display: flex;
+  gap: var(--space-4);
+  flex-wrap: wrap;
+  align-items: end;
+}
+
+.catalog-filter label {
+  display: grid;
+  gap: var(--space-1);
+  font-size: var(--step--1);
+  font-weight: 600;
+}
