+++ src/design/components/copy-button/client.ts
