+++ src/design/components/sortable-table/styles.css
@@ -1,0 +1,15 @@
+.health-table {
+  inline-size: 100%;
+  max-inline-size: none;
+  border-collapse: collapse;
+}
+
+.health-table th, .health-table td {
+  text-align: start;
+  padding: var(--space-2) var(--space-3);
+  border-block-end: 1px solid var(--color-surface-alt);
+}
+
+.health-cell--pass { color: var(--color-pass); }
+.health-cell--warn { color: var(--color-warn); }
+.health-cell--fail { color: var(--color-fail); }
