+++ src/design/components/standards-list/styles.css
@@ -1,0 +1,15 @@
+.standards-list { display: grid; gap: var(--space-2); }
+
+.standards-list__item--pass,
+.standards-list__item--warn,
+.standards-list__item--fail {
+  display: flex;
+  justify-content: space-between;
+  padding: var(--space-2) var(--space-3);
+  border-radius: var(--radius-sm);
+  border-inline-start: 3px solid var(--color-surface-alt);
+}
+
+.standards-list__item--pass { border-inline-start-color: var(--color-pass); }
+.standards-list__item--warn { border-inline-start-color: var(--color-warn); }
+.standards-list__item--fail { border-inline-start-color: var(--color-fail); }
