+++ src/design/components/tag/styles.css
@@ -1,0 +1,23 @@
+.tag {
+  display: inline-block;
+  padding: 0.1em 0.5em;
+  border-radius: var(--radius-sm);
+  font-size: var(--step--1);
+  font-weight: 600;
+  background: var(--color-surface-alt);
+  color: var(--color-ink);
+
+  &[data-variant="tier-active"] { background: var(--color-tier-active); color: var(--color-snow); }
+  &[data-variant="tier-as-is"] { background: var(--color-tier-as-is); color: var(--color-snow); }
+  &[data-variant="tier-archived"] { background: var(--color-tier-archived); color: var(--color-snow); }
+  &[data-variant="tier-unreviewed"] { background: var(--color-tier-unreviewed); color: var(--color-ink); }
+  &[data-variant="category-product"] { background: var(--color-sky); color: var(--color-midnight); }
+  &[data-variant="category-tool"] { background: var(--color-lilac); color: var(--color-midnight); }
+  &[data-variant="category-workshop"] { background: var(--color-butter); color: var(--color-midnight); }
+  &[data-variant="category-prototype"] { background: var(--color-melon); color: var(--color-midnight); }
+  &[data-variant="category-fork"] { background: var(--color-surface-alt); color: var(--color-ink-subtle); }
+  &[data-variant="category-uncategorized"] { background: var(--color-surface-alt); color: var(--color-ink-subtle); }
+  &[data-variant="pass"] { background: var(--color-pass); color: var(--color-snow); }
+  &[data-variant="warn"] { background: var(--color-warn); color: var(--color-snow); }
+  &[data-variant="fail"] { background: var(--color-fail); color: var(--color-snow); }
+}
