+++ src/design/index.css
@@ -1,0 +1,22 @@
+@layer reset, tokens, base, compositions, layout, components, utilities;
+
+@import url("./reset.css") layer(reset);
+@import url("./tokens.css") layer(tokens);
+@import url("./base.css") layer(base);
+@import url("./compositions.css") layer(compositions);
+@import url("./layout.css") layer(layout);
+@import url("./components/tag/styles.css") layer(components);
+@import url("./components/button/styles.css") layer(components);
+@import url("./components/link/styles.css") layer(components);
+@import url("./components/select/styles.css") layer(components);
+@import url("./components/card/styles.css") layer(components);
+@import url("./components/header/styles.css") layer(components);
+@import url("./components/footer/styles.css") layer(components);
+@import url("./components/repo-card/styles.css") layer(components);
+@import url("./components/standards-list/styles.css") layer(components);
+@import url("./components/catalog-filter/styles.css") layer(components);
+@import url("./components/sortable-table/styles.css") layer(components);
+@import url("./components/copy-button/styles.css") layer(components);
+@import url("./components/side-nav/styles.css") layer(components);
+@import url("./common/content-page.css") layer(components);
+@import url("./utilities.css") layer(utilities);
