+++ src/design/register.ts
@@ -1,0 +1,3 @@
+import './components/catalog-filter/client'
+import './components/sortable-table/client'
+import './components/copy-button/client'
