+++ src/pages/about.tsx
@@ -1,5 +1,5 @@
-import { ContentPage } from '../common/content-page'
-import type { SiteConfig } from '../../build/config'
+import { ContentPage } from '../design/common/content-page'
+import type { SiteConfig } from '../build/config'
 
 export function About({ body, config }: { body: string; config: SiteConfig }) {
   return <ContentPage title="About" body={body} config={config} />
