+++ src/pages/work/index.tsx
@@ -1,7 +1,7 @@
-import { Layout } from '../../common/layout'
-import { RepoCard } from '../../components/repo-card'
-import type { Catalog, CatalogEntry } from '../../../catalog/types'
-import type { SiteConfig } from '../../../build/config'
+import { Layout } from '../../design/common/layout'
+import { RepoCard } from '../../design/components/repo-card'
+import type { Catalog, CatalogEntry } from '../../catalog/types'
+import type { SiteConfig } from '../../build/config'
 
 declare module 'hono/jsx' {
   namespace JSX {
