--- src/web/components/register.ts
@@ -1,3 +1,0 @@
-import './catalog-filter'
-import './sortable-table'
-import './copy-button'
