--- src/web/styles/index.css
@@ -1,8 +1,0 @@
-@layer reset, tokens, base, layout, components, utilities;
-
-@import url("./reset.css") layer(reset);
-@import url("./tokens.css") layer(tokens);
-@import url("./base.css") layer(base);
-@import url("./layout.css") layer(layout);
-@import url("./components.css") layer(components);
-@import url("./utilities.css") layer(utilities);
