+++ tests/views/health.test.tsx
@@ -1,6 +1,6 @@
 import { describe, test, expect } from 'bun:test'
 import { renderToHtml } from '../../src/build/render'
-import { Health } from '../../src/web/pages/work/health'
+import { Health } from '../../src/pages/work/health'
 import { fixtureCatalog, fixtureNow } from '../fixtures/catalog'
 
 const config = { basePath: '/', buildTime: '2026-04-27T12:00:00Z' }
